+++ include/vinecopulib/vinecop/rvine_structure.hpp
@@ -8,7 +8,7 @@
 
 #include <Eigen/Dense>
 #include <limits>
-#include <vinecopulib/misc/nlohmann_json.hpp>
+#include <vinecopulib/misc/tools_serialization.hpp>
 #include <vinecopulib/misc/triangular_array.hpp>
 
 namespace vinecopulib {
@@ -108,10 +108,15 @@ public:
                  bool natural_order = false,
                  bool check = true);
   explicit RVineStructure(const std::string& filename, const bool check = true);
+  explicit RVineStructure(const std::string& filename,
+                          const SerializationFormat format,
+                          const bool check = true);
   explicit RVineStructure(const nlohmann::json& input, const bool check = true);
 
   nlohmann::json to_json() const;
   void to_file(const std::string& filename) const;
+  void to_file(const std::string& filename,
+               const SerializationFormat format) const;
 
   //! @return the dimension of the vine (number of variables).
   size_t get_dim() const;