+++ test/test_all.cpp
@@ -10,6 +10,7 @@
 #include "src_test/include/test_bicop_select.hpp"
 #include "src_test/include/test_discrete.hpp"
 #include "src_test/include/test_rvine_structure.hpp"
+#include "src_test/include/test_serialization.hpp"
 #include "src_test/include/test_tools_bobyqa.hpp"
 #include "src_test/include/test_tools_stats.hpp"
 #include "src_test/include/test_vinecop_class.hpp"
@@ -22,6 +23,7 @@ using namespace test_bicop_kernel;
 using namespace test_bicop_select;
 using namespace test_discrete;
 using namespace test_rvine_structure;
+using namespace test_serialization;
 using namespace test_tools_bobyqa;
 using namespace test_tools_stats;
 using namespace test_vinecop_class;